better handling of client-side closures
This required a new modification to the Nginx tree - basically it was
ignoring half-closed connections if the upstream was not yet present. I
don't totally understand this logic, because in the unhacked upstream module
should not have connections without upstreams. Nevertheless, the change is
commenting out
    if (u->peer.connection == NULL) {
      return;
    }
from the function ngx_http_upstream_check_broken_connection.

With this hack broken connections which are sitting in the queue are
properly detected and deleted.

What I don't like about this is the forced timeout that I put on the backend
slot when a connection is broken - but I can't seem to find a way to get a
callback when the upstream connection is closed. I will try to fix this now.