add ClosingBackend to test suite; fix backend close bug
The problem that has been made apparent by the new test happens when all
backends fail and a new request comes in. Because no backend is available it
is put on the queue. There is no callback that comes from a backend becoming
available again, no timeout on the queued requests, and thus the system
hangs until a new request is received.

There is not yet a test/test_handling_closing_backend.rb script - this will
come soon.