Make queue_timeout accessible from nginx.conf
I had to do this in a rather hacky way by calling ngx_parse_time() directly.
There is actually a nice way to set fields using ngx_conf_set_msec_slot()
but somehow I couldn't get this to work. It might be because this is an
upstream module?  I'll try to make that less hacky in the future.

Default queue_timeout is still 5000ms but will probably change to something
lower. What is HAProxy's timeout?