Add forced backend connections for failed requests.
Sometimes requests are dispatched to backends which end up being down. The
request returns to the peer_free function and module has the job of sending
it to a new backend. Either we drop this request (bad) or we send it to one
of the backends which already might have full maxconn. (Or put it back on
the queue - but we can't do that without hacking nginx a bit more. Perhaps
we can do this in future versions.) So this change adds some code to force a
connection to a backend in the case that one of the other backends goes down
and requests from her still need to be handled.